A special version of LineageTreeLayout for long edges.
A version of LineageTreeLayout made specially for branch graphs.

LineageTreeLayout generates Screen objects only for the vertices
that are in the visible rectangle specified in the transform when
calling cropAndScale(). This is ok when the TrackSchemeGraph is
dense, as in for the core graph, but not for the branch graph.
Indeed, the branch graph has few vertices and long edges. When
a branch does not have its vertices on the visible rectangle, it
is not painted. This version solves this issue by forcing the
generation of Screen objects ignoring the Y bounds of the visible
rectangle.