WIP: Refactor the table view.
Use a builder pattern to add content from possibly multiple graphs,
with or without some model components (selection, tags, etc).
This will be useful to display the branch graph features in the same
window.
Also this considerably lowers the constraint on the graph types
that can be displayed in such a table.