Tag-set model adapter for the branch graph.
The branch graph does not have a proper tag-set model, but mirrors
and adapt the one of the core graph.
- Setting a tag of a branch vertex sets the linked core vertex with
the same tag.
- And vice versa.
- Setting the tag of a branch edge sets this tag to all the core edges
and core vertices of the branch.
- Similarly, a branch edge *has* a certain tag iif all the core edges
and core vertices of the branch have this same tag.
This is similary to what is done to adapt the selection model to the
branch graph, so that the user has a common feeling for the model
components of the branch graph.