A common class for branch graph adapter.
This class offers two methods isValid() to test whether the edges
and vertices a branch vertex or edge are linked to are still valid.
We need to do that for when the branch graph is not in sync with the
core graph.
When the 2 graphs are not in sync,  the component models that inherit
from this abstract class should gracefully fail to do their bidingbut
not crash.