A basic coloring model.
A basic ColoringModel for a graph. The only feature color mode
supported are those based on the features defined on the vertices
and edges of the graph. The modes that rely on the branch graph are
ignored.
This is actually the former ColoringModel class by Tobias that existed
before this work on the branch graph.