Fix a bug in branch trackscheme.
The graph we paint only changes via graphRebuilt() events that
totally wipe out edges and vertices. But we still signal these
events via the graphChanged() interface.
So we need to tell the TrackSchemeGraph that it needs to be
rebuilt as well, before triggering its painting.
Because we do not report incremental changes in the branch-graph,
we don't have the nice animation in TrackScheme. This is life.