Have a set of basic feature permanently in the mamut model.
Things like position, etc. that are cheap to compute on the fly
and for which a computer is overkill. Plus having them without
truggering a computation is nice.

Also fewer classes since we can trash their computers.

Now I have to find a way to serialize them because they would be
nice for 3rd party loaders of the project files, since there are
some people that do that (MATLAB and Python).