Improve min radius when editing spot.
The min radius was a hardcoded value of 1 whatever unit, regardless of
the pixel size. This caused problems on images where the pixel size
was much smaller than 1 µm for instance.
Here the min radius is determined to be half of the smallest pixel
size in the image, so that spots edited manually always have a size
of at least 1 of the smallest pixel in diameter.

Fix #142