Use the branch iterator in the edit branch label action.
We also changed the caller method so that the action is
generic and does not specific to Mamut objects (we want
TrackScheme to be usable by all models, not only for the
graphs that are based on Spot, Link, BranchSpot and BranchLink).
As a general rule, the classes in the view package should not
have an import containing the 'mamut' package.