Notify label changed in TrackScheme using the via the UI thread.
Otherwise we get a race condition: sometimes the TrackScheme or
Branch TrackScheme view will crash because it tries to access the
label of a vertex that changed before changes are effective.

This commit works but I am still not sure this is a proper fix.
Also: the same changes could be made for the other listener in
TrackScheme view concermning the tag-sets (3 lines above) but I
am waiting to see the crash with my eyes to fix it (have not yet).