A feature serializer for features computed on the fly.
Some features, such as the link displacement, are computed on the fly,
and are not serialized. But they might be interesting for other
softwares that can read the mastodon file format and do not want
to recompute them.
So this serializer utility can be used to serialize a generic feature
even if it does not store a map of values.