A specialized FeatureSerializer for the branch features.
The de/serialization process is a bit complex in the case
of branch features. We serialize only the core model, not
the branch graph. So we get access to a map the maps the
core graph object (spots ot links) to file ID. But the
features are defined over a branch spot or a branch link.
We we need this interface that invalidates the mother
interface (calling the non-specialized methods will
trigger a IllegalArgumentException) and can deal with
the issue.