PaintGraph: don't render text that isn't on the screen anyway
It happened that labels were randomly drawn in wrong places in the
TrackScheme Branch window. The likely cause is that labels that shouldn't
have been visible because they were far out of view were misplaced and
rendered on screen instead. The incorrect placement may be due to a number
overflow in the AWT library.

The issue is addressed by testing whether a vertex center is in the
visible area on the screen. And the label will only be drawn if this
condition is met. This avoids drawing text far outside the visible area.