Fix: prevent TrackScheme Hierarchy from jamming the TimepointModel
There was a bug, when first linking a normal TrackScheme and a BDV view,
and afterwards linking a "TrackScheme Hierarchy" to it. Clicking a node
in a TrackScheme window would navigate to the wrong time point in the BDV
window. This bug was caused by the Hierarchy TrackSchemePanel. The Hierarchy
TrackSchemePanel works with pseudo time points, and wrote these pseudo
time points into a global TimepointModel. BDV would as a consequence show
the pseudo time points.

This problem is fixed by providing the hierarchy TrackSchemePanel
with its very own "detached" TimepointModel.