Fix column titles in the TrackScheme Branch
Change the column titles in the TrackScheme Branch to match the
normal TrackScheme.

It was necessary to add a getFirstLabel() to the ModelGraphProperties, this
method return the label of the first spot represented by a branch spot.
getFirstLabel is used for column titles and column sorting to be consistent
with the normal TrackScheme.