Fix SelectedTracksActions.showTrackDownward() when used on root node.
Fix a bug triggering a NullPointerException when showTrackDownward()
is used with a root node of the TrackScheme selected.

The problem was caused by the DepthFirstIteration class. The truncate()
method didn't properly end the iteration, when being called on the
root node. It iteration would continue instead, and later trigger a
NullPointerException.

This commit fixes DepthFirstIteration.truncate() to work properly
on the root node.