Better copy from vertex settings in feature color mode.
With the new branch graph painted in TrackScheme, if we want to achieve
a coloring that is coherent between the normal TrackScheme and the
branch TrackScheme, it is best to  assign a branch link color from
the color if its TARGET branch spot. So let's make this the default.
Otherwise, the links are represented by two colors.