Change how TrackSchemePanel calls TimepointModel.setTimepoint(...)
The old behaviour for setting the timepoint model in the method
"TrackSchemePanel.navigateToVertex" had some strange implications.
The timepoint model is set as a response to a navigate-to-vertex-event,
but only if there was a TrackSchemePanel open. And several open
TrackSchemePanels led to the timepoint model to being set several times.

With the new "TrackScheme Branch" window we get wrong behavoir.
The "TrackScheme Branch" window sets the wrong timepoint. As the
navigate-to-vertex-event is adapted to branch-spots. And the branch-spot's
timepoint might not match the original (navigate-to-vertex) timepoint.

This commit introduces a consistent behaviour to set the timepoint model.
The timepoint is set exactly once if a spot is focused by the user in
the active TrackSchemePanel. Focusing a spot in the active
TrackSchemePanel will call the method focusVertex(vertex) of the
AutoNavigateFocusModel. And this method calls:
1. TimepointModel.setTimepoint(vertex.getTimepoint)
2. FocusModel.focusVerex(vertex)
3. Navigationhandle.navigateToVertex(vertex)