Fix some menu items being inactive in the BDV views.
The menu builder must be called apparently AFTER the corresponding
actions have been installed. Lest these actions are inactive when
called from the menu (but this works with a shortcut).

This requires some gymnastic, because the coloring requires the
menu handles to be built before. So we declare those early in the
initialization process. And to keep the future menu
order, we build an empty menu but already with all su-bmenus in
order.
Then at the end of the initialization process we flesh out all the
menu items.