Change MamutProjectIO to correctly use Windows file paths on Linux
Windows, Linux & MacOS use different slashes in file paths. Windows
uses backward slashes, Linux & MacOS use forward slashes.

If a Mastodon project is saved on Windows, the project.xml will have
a dataset path with backward slashes like: ..\dataset.xml.
On Linux this path is not interpreted correctly. We need to manually
replace the backward slashes with forward slashes.

This commit moves the code that replaces the slashes into the class
MamutProjectIO. This makes a lot of sense, as the backward slashes
are introduced when reading the project.xml.