ShowSelectedTracksAction: don't get stuck if there's a loop in the graph
The "show track downward" action got stuck, whenever there's a loop in the
ModelGraph. This is because the "show track downward" action uses the
DepthFirstIteration class to iterate over the ModelGraph when searching
for the correct new roots.

The DepthFirstIteration class itself does not detect loops and hangs if
loops are present in the graph. The commit adds the required tests to
that keep track of visited nodes and breaks the loop during iteration.