TreeUtils: change semantics of findRootsOfTheGivenNodes
The method findRootsOfTheGivenNodes is used in the
ShowSelectedTracksAction. It's purpose is to find the root nodes of the
tracks that contain selected nodes. The new implementation gives better
results if the ModelGraph also contains merges.

The previous implementataion worked fine if the ModelGraph contained only
trees. But it did miss root nodes for more general graphs. In order to
get all roots it is necessary to do a depth first search that considers the
graph as undirected.

Example graph:
a   b
| \ |
c   d

If "c" is selected, the method now return both roots "a" and "b" of the
track.