Example of a manual branch feature.
We want to build a branch feature that can be set 'manually', that
is without using the feature computation framework. But we still
want to use it everywhere in Mastodon, and have it properly managed
and cleaned when a user delete a branch.

The example abstract class BranchScalarFeature does that for scalar
features with double values.

Normally we could simply build such
a feature with the DoubleScalarFeature class, but the issue is that
branch objects are 'volatile': they all get destroyed and recreated
everytime an edit is made to the model (and the branch graph is
refreshed). This leads the manual branch feature values to be
forgotten everytime this happens.

The solution proposed here is to actually store feature values on
the last spot of the branch. The value is access through the branch
but is defered to the spot item. This way we retain the branch feature
values against multiple changes.

There is always a limitation with invalidating feature values. Here,
the branch feature values are invalidated only if the last spot of the
branch is modifed (moved or deleted). This is something we will document.

TODO: Now we can also offer a de/serializer for such a feature, so that
its values are persisted.