DummyGraph: fix memory leak, the selection model must not be "static"
Examples.CELEGANS is never garbage collected because it's a static
constant. This is not a problem at itself. But it means that
Examples.CELEGANS.selectionModel is never garbage collected.
And also all the classes that listen to the selectionModel
are never garbage collected. Which in that case is indirectly the
preferences dialog. And the preferences dialog indirectly references the
appModel. As a consequence all Mastodon data structures are never
garbage collected.

This commit solves this problem by removing the field
Examples.CELEGANS.selectionModel.