Rework project loading / saving.
This branch reflects an effort to rework the main class of the Mamut
app: the MamutAppModel. The main goal is to make the app model the
core of a Mastodon session, and have all of its fields final, and
have them containing everything we need to run all featuures of the
app.

This is a WIP but so far we have:
- The windowManager is now a field of the app model.
- The app model is immutable: it cannot change under the user feet.
When a user loads a project, it gets another instance, but cannot
modify an existing one.
- And to do this, the main work of this PR is the rework of the project
loading and saving, detailed a bit below.

The ProjectManager class is gone, replaced by 4 classes with only static
methods, in the package org.mastodon.mamut.io. Each one takes care
about loading, saving, importing or exporting.

The loading methods all produces a new MamutAppModel instance. This
instance receives the MamutProject it was created on (handy to save
it later). The app model is properly initiatialized ONCE with
everything. Then a consumer can create a MainWindow to pilot it.

I have updated all the classes so that they compile. The tests all
run (except for the garbage collection from Mathias, which requires
changing two deps). But I am anxious to see if we cover all the use
cases below:

- [x] open a project
- [ ] save a modified project
- [x] 'save as' a project
- [ ] test that we can reload a saved project
- [ ] test that we can reload a 'saved as' project

The importers and exporters work, as seen by the tests, but I would like
to test them and rework their UI.
Then I will rework again the app model, rename it etc...