Store style managers in a common map.
Mainly as an effort to reduce the number of fields and methods
in the window manager class. We began with just a manager for BDV
and TrackScheme styles but now we have 5 of them. For now.
Also this will make it easier to extend Mastodon with
new views in the future.