Manage views and branch views in a common fields.
The view lists are stored in a map, indexed by their class, like for
the manager. This diminish the number of fields and methods in the
window manager class.

This also allows for treating views in a general manner without
having to change too much code when a new view will be added.
For this I shall find a way to create view instances in a generic
manner, probably with a factory pattern.