Tentative factory design for mamut views.
We started with just 2 views (BDV and TrackScheme) but now there are
enough of them is that it is worthwhile having some generic way of
handling them. The next commits will be about coming with a sensible
way of doing it. Ideally, we want to make it possible for 3rd parties
to extends Mastodon with new views without having to recompile it.