Use the new view factories everywhere in Mastodon.
The window manager has now a MamutViews that is in charge of
detecting and managing a collection of view factories.

Each factory is then used by the window manager to make one action
that creates a view, registers it in the WM, deals whether it has a
context provider or a context chooser, notifies listeners of its
creation (see below) and finally shows it.

New views are created, registered and managed only via these factories
now. Which means that that window manager does not know of any
concrete view class. It just manages all the ones it found. This
will make it easier to extends Mastodon with new views, without
having to touch or even recompile the core.

This also makes the window manager class much shorter. There are much
less methods and much less fields, since all views are managed in a
generic manner.

The view factory are also nice because they contain in one place all
the info related to an action: command name, menu text, description.

For ELEPHANT, Ko requires that there is a view creation listeners list
for BDV window. I kept this, but made it generic. There is now one
listener list per view type. So you can register a listener for the
creation of any type of view, and get it befoore it is shown.

The factories also handle GUI state serialization. The ProjectLoader
and ProjectSaver class use these factories when loading / saving for
restoring / saving the GUI state. There are no specific hard-coded
code to handle serialization of GUI state in the io package. Well,
not entirely: we still need to know how to convert a Map< String, Object >
(which is the GUI state as understood by the views) to a JDom element,
which is what the MamutViewStateXMLSerialization does. It just has
a series of methods to convert any type or class used in GUI state
to a JDom-ready text.

This de/serialization also now takes much less code than before, as
there were a lot of duplicated code, wich is now present once in
the mother abstract class for view factories.

Also: the views are now decoupled from saving or restoring the GUI
state. The factory does it, which is cleaner.

The MainWindow class still needs to know of concrete view types,
because we still want to have buttons for the views, and have the nice
click and shift-click on them. I think this is ok
because normally the rest of the code is *almost* fully decoupled
from the MainWindow. There is still one static method addMenus()
that is called by the views to create the File menu. This could
be moved to the window manager, as it is as generic as the Window
menu.

This will make it easier also, to create new Mastodon commands or
plugins that do now show the MainWindow, but whatever a 3rd party
coder wants.