Use a common utility to discover plugins.
There is the pluginService.createInstancesOfType() that is very
concise, but I want to make it possible for a plugin to be
disabled, using the flat Plugin( type = ....class, enabled = false ).
In that case and with the new utility, the class is not instantiated
by the service, which makes it possible to use it "manually", or
disable unwanted plugin simply.