Replace KeyMap and associated classes by upstream implementations.
These classes have been migrated in upstream repos, mainly the
BDV-core and ui-behaviour. We can remove them from Mastodon and use
the common ones, maintained elsewhere.

There are two adaptations specific for Mastodon:

1/ The KeymapManager class is extended by a specific one for Mastodon
so that we can load and save keymaps in the Mastodon folder, and so
that we load the builtin keymaps we had before.

2/ The new CommandDescription requires specifying a context AND now a
scope. I made two scopes: one for the actions that are generic to
Mastodon (everything in views, that do not know of the Mamut model),
and one specific for the Mamut application (everything under
org.mastodon.mamut, which is the one users have).

I also made a static method to create a ProjectModel, and hide the
creation of the KeymapManager and KeyPressedManager there.

I could not replace one class:
- KeymapSettingsPage: There is the equivalent in BDV-core, but it
does not 'see' the builtin keymaps we have in Mastodon.