Recompute min & max timepoint everytime they are queried.
During a discussion with @stefanhahmann
We don't need to store the values determined at startup, because
we can get them from the current shared BDV data. In theory this
makes it possible to have Mastodon views reflects changes in the
image data (even if we have no cases for it yet). Plus, it's shorter
and more elegant.
However, we will need to double-chek that everything is ok with
BDV files, ImagePlus and DUMMY image file format.