Offer facilities to deserialize from YAML with catchable errors.
Ko made me realize that with slight deviation from what is expected
in a YAML file, Mastodon will crash with no recoverable exception.
This commmit introduces some methods to deserialize common core
values that gracefully trigger an exception, with a useful error
message. This exception is caught by the load() method of the style
manager, which handles them gracefully.