Remove superficious type in the FocusModel.
The FocusModel interface had an unneeded type. It does not need to
know the type of edges in a graph, because it does not work with a
graph, but with a simple collection.
Removing this superficious type triggered a lot of changes in the
code because we touched the signature of a core interface.
But not impact otherwise.