When restoring a BDV view, wait a bit before restoring the transform.
If we don't wait a little bit, the BDV state, notably the
transform, is not restored properly. It is not enough to
just display the frame before restoring the transform.

We put the delay 0.1s in a separate thread not to block
the loading if we have many BDV views.

Fixes #251