A color generator that gives one color per track.
It works like this:
- First all the roots of a graph are collected.
- They are sorted alpha-numerically, so that they are iterated in
the same order than for default TrackScheme (with no fusion points).
- Each root is given a color from the Glasbey LUT, in order. This
ensures that all track color generators will give the same color
for one vertex.
- When the color for a vertex is quiered, we look in a cache. If the
color is not in the cache, then we iterate 'upward' in the graph
(reverse depth-first order) until we meet a spot that has a color
in the cache. All the spots we iterated until then receives this
color in the cache.
- For edges, we quiery the color of the source vertex.

The color generator listens to graph changes, so that new colors are
generated when the user adds a new track or split one in two.
Its de/registration as a listener must be handled outside.
Listening to graph changes can be paused.