A class that listens to a graph to provide the set if its roots.
Similar to RootFinder, except that it tries not to walk the full
graph everytime it is queried for the roots.

The root set is updated by listening to graph changes. So to work,
aninstance of this class must be first registered as a GraphListener
to the graph it will analyze, and then have its graphRebuilt() method
called once.

If this works well, this should be transferred to mastodon-graph,
in the package org.mastodon.graph.algorithm