Rework the track color generator for better performance.
I foresee that it will not be good to generate the track colors,
by waling the full graph ONCE per opened view TIMES ONCE per
change in the graph.

So this version:

- Is based on the RootProvider, and listens to it to only regenerate
the cache when the set of roots changes.

- There is only one common instance of the track color generator,
use by all views, and it is registered as a manager in the window
manager. I took some care of what happens if several threads try to
query the track color concurrently, at least partially.

- The color cache is regenerated only if there is a change in the
root set AND if at least one view is using the 'Per track' coloring
mode.