FeatureSpecsService must have EXTREMELY_LOW priority.
Doing otherwise causes a very weird bug, making Mastodon crash when
deployed in Fiji but not in Eclipse.

The reason for the crash is the following: When creating the
FeatureSpecsService, it immediately runs Specs discovery. For this
it needs the PluginService to be ALREADY initialized. This is the
case when running Mastodon from Eclipse, but not when running it
from Fiji.
Setting the priority to extremely low makes sure FeatureSpecsService
is initialized among the last, well after the services it depends on.