Make sure the projections() method of feature is called only once,
when creating the table header and collecting values.
This is to protect against a bug that appear with features
that are not careful to always return the same set of
projections when the projections() method is called. This
happend for the link target feature and generated the bug below:
https://github.com/mastodon-sc/mastodon/issues/217

One can fix the bug by imposing the same order on features, but
it is also a plus to make the table rendering independent of
this order being preserved.