Let the MastodonLogger have a subLogger method
This way to user does not have to deal with LogSource. Instead of creating
a LogSource.subSource(...) simple use logger.subLogger(...) this will
create a subLogger that uses a "sub source".

Methods like setStatus(message, source) become obsolete.

There is now instead the problem, to pass the proper sub-logger to a given
plugin.