Work around (manual) features which specs change every calculation.
The best example is the NN stat feature. Depending on what the user
chooses to compute, the spec of the NN feature will contain a list
of *projection specs* that changes every time. But is still ok to
require these projections.
So here, when we are tasked with retrieving a specific projection value,
we simply get the feature, and read the spec from the feature we just
collected instead of the one the user had before.