Replace custom distance method by distance method of imglib2
While searching for a method that can directly compute the distance between 2 spots in the Mastodon code base, I found the distance method in this class. Later, I learned that there is the same method in imglib2.

In order to simplify the code I suggest to replace this re-implementation by using the imglib2 version of this method.