SpotPool: increase capacity, allow more than 23 million spots in Mastodon
Using the MultiArrayMemPool rather than SingleArrayMemPool eliminates the
upper bound for the number of Spots in Mastodon. Without this change
Mastodon is limited to approx. 23.3 million spots. Reason for that is that
the SingleArrayMemPool limits the available space for spots to 2 GB.
MultiArrayMemPool removes this limitation.

The are other similar pools in the source code, that probably would need to
be changed to avoid similar problems these are at least:
* LinkPool
* DataGraph
* TrackSchemeGraph