A plugin to export the current model to a MaMuT file.
So that we can use the tools made for TrackMate file as well.
It's funny, we are using TrackMate XML format as an interchange
file format. It is not well suited for this I think, but waiting
for such a file format to emerge, we will do with what we have.