RootsModel: prevent exception that occurs when removing a root node
There was a bug, that could be reproduced as follows:
Open a Mastodon project with multiple tracks.
Select a few tracks and click "View > Show Only Selected Tracks".
Remove one of the tracks.
Result: TrackScheme window would freeze and the console would show
NullPointerException.

The bug is triggered because the DefaultRootsModel and
BranchTrackSchemeRootsModel keep a list of roots to be shown in the
TrackScheme view. If one of the root is removed from the ModelGraph, it
becomes invalid, but the invalid entry still remains in the list of roots.
The exception is triggered when the invalid object is processed later on.

This commit fixes this bug, by removing the a entry fro the list of roots
as soon the corresponding Spot is removed from the ModelGraph.