Have a global behaviour to add or link to a new or existing spot.
This new behaviour tries to encompass all what the 'add spot' and
'add linked spot' do.
If 'A' (the default key binding) is pressed in a zon without a spot,
a new spot is created. If it is pressed within a spot, the viewer
moves to the next frame ('C' to move to the previous one) and the user
can drag around to create a new linked spot there. If the key is
released in a zone without any spots, a new one is created, from the
source spot. If a spot is found near where the mouse is released,
the source spot is linked to this one.