More stable track coloring.
Prior to this commit, the track color was determine by the position
of a track in the sorted track array, based on the root labels.
In practice, for large number of tracks, this was not so great when
editing tracks a lot. Indeed, the slightest track additon of link
removal could completely change the order, hence the track coloring.
Here we simply use the hash of the root label of a track to determine
its color, making it independent of the other ones. The color of a track
now only depends on the label of its root.