Avoid duplicated attributes when export to MaMuT a file imported from MaMuT.
During an export to MaMuT, the position of the spots are encoded in the
feature 'POSITION_X' etc, so that MaMuT can retrieve the spots at the
right position.

A problem might occur if the data in Mastodon have been imported from
a MaMuT (or TrackMate) file. In that case the exporter will try to
write two attribute with the same name: the POSITION_X from the actual
spot position and the POSITION_X that was imported. This is happening
for all the builtin features we need for the MaMuT export.

To avoid this, we rename the imported features with 'IMPORTED_...'. And
if we have imported features already named 'IMPORTED_...' (if the user
re-reimport a MaMuT exported file from a MaMut imported file), we skip
them.

Fix #364