Refactor Model and its branch.
The branch graph part of the model is managed in an inner class,
with methods name that are similar to the outer model. This makes
sense because the branch graph is just a graph.
In case we want to go further, the model methods could implement
an interface, that the branch model inner class could implement as
well.

This needs to be further developed in the ProjectModel.