WIP: Trying to generalize views.
For this we create a model interface, generic to Mastodon and we
try to write views that will directly accept this model.

Very tricky. MamutViewTrackScheme2 is the demo target of this.

Right now I think we still need to have a class to store app
related components (the managers, the windowmanager, the plugins,
the actions etc). Not sure.