Remove the timepoint model from the BranchGraphHighlightAdapter
This makes it a proper data model class, without any link to a ui
model. The issue is that this was nice to highlight the current
vertex in core graph view in the current timepoint. With this
change we loose this feature.
I will have to put it back one way or another, but on the adapters
that are proper to the views, not to the model.