New TrackScheme view takes dataModel and uiModel as parameters.
Instead of the whole appModel. The aim is to use the same class
to create a view of the core graph and of the branch graph. So
we cannot pass the appModel which is specific to a MastodonModel.

When we do this, we will have to deal with UndoPointRecorder.
Probably an intermediate class will be needed in cases where the
graph will not be editable in the TrackScheme view.