WIP: still struggling with irreconciliable hierarchies.
According to Claude:

"The class MamutViewTrackSchemeFactory2 is attempting to implement
the MastodonViewFactory interface twice with different type
parameters:

Once through extending MastodonViewTrackSchemeFactory<ModelGraph>,
which ultimately gives you MastodonViewFactory<MastodonViewTrackScheme2
<?,ModelGraph,?,?>>

Again through implementing MamutViewFactory2, which gives you
MastodonViewFactory<MamutView2>

Java doesn't allow a class to implement the same interface with
different type arguments - this creates an ambiguity that the
compiler cannot resolve."

None of the solution I tested worked or were elegant. I still need
to have a sub-hierarchy of factories that implement MamutViewFactory2
so that I can discover them in the Mamut app. And I still want to
extend MastodonViewTrackSchemeFactory in MamutViewTrackSchemeFactory2
because is is the way to go.

I need to think of this now.